+++ javamesh-plugins/javamesh-server-monitor/metric-server/src/main/java/com/huawei/javamesh/metricserver/dao/influxdb/entity/openjdk/YoungGCInfluxEntity.java
@@ -1,0 +1,30 @@
+/*
+ * Copyright (C) Huawei Technologies Co., Ltd. 2021-2021. All rights reserved
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.huawei.javamesh.metricserver.dao.influxdb.entity.openjdk;
+
+import com.influxdb.annotations.Measurement;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+
+/**
+ * OpenJdk jvm metric young gc Influxdb持久化实体
+ */
+@Data
+@EqualsAndHashCode(callSuper = true)
+@Measurement(name = "oracle_jvm_monitor_young_gc")
+public class YoungGCInfluxEntity extends GCInfluxEntity {
+}
