+++ javamesh-plugins/javamesh-server-monitor/metric-server/src/main/java/com/huawei/javamesh/metricserver/service/OpenJdkJvmMetricService.java
@@ -1,18 +1,30 @@
 /*
- * Copyright (c) Huawei Technologies Co., Ltd. 2021-2021. All rights reserved.
+ * Copyright (C) Huawei Technologies Co., Ltd. 2021-2021. All rights reserved
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 package com.huawei.javamesh.metricserver.service;
 
 import com.huawei.javamesh.metricserver.dao.influxdb.InfluxDao;
-import com.huawei.javamesh.metricserver.dao.influxdb.entity.skywalkingjvm.CpuInfluxEntity;
-import com.huawei.javamesh.metricserver.dao.influxdb.entity.skywalkingjvm.GCInfluxEntity;
-import com.huawei.javamesh.metricserver.dao.influxdb.entity.skywalkingjvm.OldGCInfluxEntity;
-import com.huawei.javamesh.metricserver.dao.influxdb.entity.skywalkingjvm.ThreadInfluxEntity;
-import com.huawei.javamesh.metricserver.dao.influxdb.entity.skywalkingjvm.YoungGCInfluxEntity;
-import com.huawei.javamesh.metricserver.dto.skywalkingjvm.CpuDTO;
-import com.huawei.javamesh.metricserver.dto.skywalkingjvm.GcDTO;
-import com.huawei.javamesh.metricserver.dto.skywalkingjvm.ThreadDTO;
+import com.huawei.javamesh.metricserver.dao.influxdb.entity.openjdk.CpuInfluxEntity;
+import com.huawei.javamesh.metricserver.dao.influxdb.entity.openjdk.GCInfluxEntity;
+import com.huawei.javamesh.metricserver.dao.influxdb.entity.openjdk.OldGCInfluxEntity;
+import com.huawei.javamesh.metricserver.dao.influxdb.entity.openjdk.ThreadInfluxEntity;
+import com.huawei.javamesh.metricserver.dao.influxdb.entity.openjdk.YoungGCInfluxEntity;
+import com.huawei.javamesh.metricserver.dto.openjdk.CpuDTO;
+import com.huawei.javamesh.metricserver.dto.openjdk.GcDTO;
+import com.huawei.javamesh.metricserver.dto.openjdk.ThreadDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -20,10 +32,10 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 @Service
-public class SkywalkingJvmMetricService extends InfluxService {
+public class OpenJdkJvmMetricService extends InfluxService {
 
     @Autowired
-    public SkywalkingJvmMetricService(InfluxDao influxDao) {
+    public OpenJdkJvmMetricService(InfluxDao influxDao) {
         super(influxDao);
     }
