+++ javamesh-plugins/javamesh-server-monitor/metric-server/src/main/java/com/huawei/javamesh/metricserver/service/OpenJdkMemoryService.java
@@ -1,14 +1,26 @@
 /*
- * Copyright (c) Huawei Technologies Co., Ltd. 2021-2021. All rights reserved.
+ * Copyright (C) Huawei Technologies Co., Ltd. 2021-2021. All rights reserved
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 package com.huawei.javamesh.metricserver.service;
 
 import com.huawei.javamesh.metricserver.dao.influxdb.InfluxDao;
 import com.huawei.javamesh.metricserver.dao.influxdb.entity.MemoryPoolInfluxEntity;
-import com.huawei.javamesh.metricserver.dao.influxdb.entity.skywalkingjvm.HeapMemoryInfluxEntity;
-import com.huawei.javamesh.metricserver.dao.influxdb.entity.skywalkingjvm.NonHeapMemoryInfluxEntity;
-import com.huawei.javamesh.metricserver.dto.skywalkingjvm.OracleMemoryDTO;
+import com.huawei.javamesh.metricserver.dao.influxdb.entity.openjdk.HeapMemoryInfluxEntity;
+import com.huawei.javamesh.metricserver.dao.influxdb.entity.openjdk.NonHeapMemoryInfluxEntity;
+import com.huawei.javamesh.metricserver.dto.openjdk.OpenJdkMemoryDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -16,36 +28,36 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 /**
- * Oracle Memory 服务
+ * OpenJdk Memory 服务
  */
 @Service
-public class OracleMemoryService extends InfluxService {
+public class OpenJdkMemoryService extends InfluxService {
 
     @Autowired
-    public OracleMemoryService(InfluxDao influxDao) {
+    public OpenJdkMemoryService(InfluxDao influxDao) {
         super(influxDao);
     }
 
     /**
-     * 添加{@link OracleMemoryDTO}实体
+     * 添加{@link OpenJdkMemoryDTO}实体
      *
-     * @param memoryDto 待添加的{@link OracleMemoryDTO}实体
+     * @param memoryDto 待添加的{@link OpenJdkMemoryDTO}实体
      */
-    public void addMemoryMetric(OracleMemoryDTO memoryDto) {
+    public void addMemoryMetric(OpenJdkMemoryDTO memoryDto) {
         insert(() -> newEntityOfType(memoryDto.getType()), memoryDto);
     }
 
     /**
-     * 查询{@link OracleMemoryDTO}实体列表
+     * 查询{@link OpenJdkMemoryDTO}实体列表
      *
      * @param type  Memory类型
      * @param start 开始时间
      * @param end   结束时间
      * @return @link OracleMemoryDTO}实体列表
      */
-    public List<OracleMemoryDTO> getMemoryMetrics(OracleMemoryDTO.OracleMemoryType type, String start, String end) {
+    public List<OpenJdkMemoryDTO> getMemoryMetrics(OpenJdkMemoryDTO.OracleMemoryType type, String start, String end) {
         return query(start, end, getClassOfType(type)).stream()
-            .map(entity -> OracleMemoryDTO.builder()
+            .map(entity -> OpenJdkMemoryDTO.builder()
                 .service(entity.getService())
                 .serviceInstance(entity.getServiceInstance())
                 .time(entity.getTime())
@@ -57,7 +69,7 @@ public class OracleMemoryService extends InfluxService {
             .collect(Collectors.toList());
     }
 
-    private MemoryPoolInfluxEntity newEntityOfType(OracleMemoryDTO.OracleMemoryType type) {
+    private MemoryPoolInfluxEntity newEntityOfType(OpenJdkMemoryDTO.OracleMemoryType type) {
         switch (type) {
             case HEAP:
                 return new HeapMemoryInfluxEntity();
@@ -68,7 +80,7 @@ public class OracleMemoryService extends InfluxService {
         }
     }
 
-    private Class<? extends MemoryPoolInfluxEntity> getClassOfType(OracleMemoryDTO.OracleMemoryType type) {
+    private Class<? extends MemoryPoolInfluxEntity> getClassOfType(OpenJdkMemoryDTO.OracleMemoryType type) {
         switch (type) {
             case HEAP:
                 return HeapMemoryInfluxEntity.class;
